requirements.txt
google-genai
google-cloud-texttospeech
python-dotenv
requests
Pillow
